requirements.txt
streamlit
faiss-cpu
google-generativeai
langchain
langchain-community
langchain-google-genai
python-dotenv
requests
pandas
